refactor: Move components related to laboratory edit view
Move components closer to their screen / page
