+++ src/screens/edit-laboratory/components/EditableMarkdownBlock.tsx
