+++ src/screens/edit-laboratory/components/MarkdownBlockDropDown.tsx
@@ -12,7 +12,7 @@ import {
   DropdownMenuLabel,
   DropdownMenuSeparator,
   DropdownMenuTrigger
-} from "../ui/dropdown-menu";
+} from "../../../components/ui/dropdown-menu";
 
 interface MarkdownBlockDropDown {
   blockUUID: string;
